+++ src/emw_keys.cpp
@@ -334,7 +334,7 @@ void MainWin::xkey_press(int xkey) {
 				if (comp->rzx.play) break;
 #endif
 				if (comp->cpu->type != CPU_Z80) break;
-				comp->flgNMIRQ = 1;
+				comp_irq(IRQ_NMI, comp);
 				break;
 			case XCUT_TAPWIN:
 				emit s_tape_show();